Use the same plugin class name in different map engine implementations
