+++ bundles/mapping/mapmodule/MapModuleClass.ol.js
